Online evaluation/chatbox stuff
More lines for the chatbox output, with a smaller font(Total size is the same). I'm also submitting changes to _chatbox.lua in BAnims to properly change the size of the quad/square for the background of the chatbox.

Added [ScreenNetEvaluation] and added a second set of judgements and options which update when you scroll through users when you play online. Since it's in NetEvaluation it doesn't affect regular offline play.