Keeping the screen title colortype the same
This change is made to keep this colortype and ScreenSelectMusic's colortype the
same. I want to make sure the color visibility remains one color and not
having two different colors.

Changing colortype in ScreenSelectProfile overlay

This change also have the same reason as the commit above.

Cleaning ScreenNetSelectMusic song sort to look like offline

This commit focuses on cleaning up comments and removing quads in song
sort because it's not needed anymore. I'm not in the mood of having it
look like purple recolor of Prim's spawnhack, so I decide to change it.

Cleaning up song sort to look like offline

This also have the same reason as the previous explanation.

Comment removed and cleaned up

This commit is made so people don't have to see this mess when they open
up this file.