Remove white quad from SMO select music screen
I just found out I messed up the song sort in ScreenNetSelectMusic title overlay because I didn't check it 100%, therefore this should resolve the issue. The issue is it's showing the white quad when I want it to show the song pack title in the same font size as the offline one.